fix: Feedback cumulativeElapsedTime 타입을 Integer에서 Long으로 변경 (#45)

+++ src/main/java/com/cotato/blankit/domain/feedback/service/FeedbackService.java
@@ -109,7 +109,7 @@ public class FeedbackService {
             intervalDiff = (int) B;
         }
 
-        feedback.updateMetrics(prevRate, (int) cumulativeElapsedSeconds, consecutiveCount, intervalDiff);
+        feedback.updateMetrics(prevRate, cumulativeElapsedSeconds, consecutiveCount, intervalDiff);
     }
 
     private TaskSession getSessionAndVerifyOwner(Long userId, Long sessionId) {
